+++ src/components/templates/ChatRoomTemplate/TopSheet.tsx
@@ -93,11 +93,10 @@ export const TopSheet = ({ post, isCompleted }: TopSheetProps) => {
         isOpen={isOpen}
         isSeller={post.isSeller || false}
       >
-        {isOpen
-          ? ""
-          : <Text variant="tag_regular">
-              {post.title}
-            </Text>}
+        {!isOpen &&
+          <Text variant="tag_regular">
+            {post.title}
+          </Text>}
         <div className="btn-bar" />
       </ToggleButton>
       <TopSheetContainer isOpen={isOpen} isSeller={post.isSeller || false}>
