+++ src/types/post.d.ts
@@ -43,3 +43,39 @@ export interface ILocation {
   coord?: ICoord;
   address?: string;
 }
+
+export interface IPost {
+  /** 게시글 ID */
+  productId: number;
+  /** 게시글 썸네일 이미지 */
+  imgUrl: string;
+  /** 게시글 제목 */
+  title: string;
+  /** 게시글 가격 */
+  price: number;
+  /** 게시글 등록된 주소 */
+  address: string;
+  /** 게시글 등록된 시간 */
+  uploadTime: string;
+  /** 게시글 아이템 클릭 이벤트 */
+  onClick: () => void;
+  /** 남은 시간 */
+  expiredTime: string;
+  /** 판매자 : 최고 입찰가, 구매자 : 나의 입찰가  */
+  maxPrice: number;
+  /** 판매중 : 끌어올리기, 완료 : 받은 후기 보기  */
+  onTextButtonClick: () => void;
+  /** 아이콘 버튼 클릭 이벤트 */
+  onIconButtonClick: () => void;
+  /** 판매자 여부 */
+  isSeller?: boolean;
+  /** 게시글 상태 */
+  status?: "BIDDING" | "IN_PROGRESS" | "COMPLETED";
+}
+
+export type PostItemType =
+  | "completed"
+  | "default"
+  | "chat"
+  | "selling"
+  | "buying";
