+++ src/types/response/index.ts
@@ -10,3 +10,4 @@ export * from "./areaAuth";
 export * from "./transaction";
 export * from "./block";
 export * from "./post";
+export * from "./chat";
