+++ src/types/response/post.d.ts
@@ -1,4 +1,3 @@
-import { IResponse } from "types";
 
 export interface ISearchPost {
   myLocation: string;
